Add of cute library to the project template
CMakeLists.txt reworked for better reading.
It also contain, the beginning for a superbuild option.

linux-clang.cmake modified because warnings flags of
GNU gcc aren't for most of them understood by clang.

simpletest.cpp is a file copied from the cute library,
used to test the correct linking to the library.

Signed-off-by: Guillaume Jeusel <jeusel.guillaume@gmail.com>